update module name keys -> credentials
